po-mode: make (po-check-file-header) optional (bug #30835)
po-mode would always overwrite the PO Header Entry when editing *.po
files. This isn't always desired behavior as noted in bug #30835 so
it's now customizable.

It can be customized through the po-auto-update-file-header variable,
which can be t, nil or 'ask. It's t by default to preserve the old
default behavior.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>