build: enable parallel tests harness from Automake
This way, we'll be able to run test cases in parallel (useful
on multicore systems), and output from test cases will be
saved in log files, which should simplify debugging and bug
reporting.

* configure.ac (AM_INIT_AUTOMAKE): Add 'parallel-tests' option.
Require Automake 1.11.1 or later, so that the 'parallel-tests'
will be certainly available.
(AC_INIT): Require Autoconf 2.62 or later; that is the minimal
version supported by Automake 1.11.1 or later.
* gettext-runtime/configure.ac: Likewise.
* gettext-tools/configure.ac: Likewise.
* autogen.sh: Update version number requirement for Autoconf.
* gettext-tools/tests/Makefile.am (TESTS_ENVIRONMENT): Remove
trailing '$(SHELL)'.  With the parallel-tests harness, this is
no longer the correct way to define a custom test runner for
the test scripts; to do so, we have to ...
(LOG_COMPILER): ... define this to $(SHELL).

Copyright-paperwork-exempt: yes
Signed-off-by: Stefano Lattarini <stefano.lattarini@gmail.com>