Add support for msgmerge --previous
msgmerge --previous is a very useful feature that makes adjusting of
translation much easier when small changes in the source code are done.
msgmerge supports it for 10 years, but it is not used by most projects,
as Makefile.in.in never added support for it.

Use msgmerge --previous as default on all systems with gettext >= 0.16.