examples: hello-java*: Fix "make dist".
* gettext-runtime/m4/po.m4 (AM_POSTPROCESS_PO_MAKEFILE): Change PROPERTIESFILES
and CLASSFILES to contain files in $(srcdir), not in $(top_srcdir).
* gettext-tools/examples/hello-java*/Makefile.am (hello.jar): Fetch the catalogs
from the po/ directory.
* gettext-tools/examples/hello-java*/po/Makefile.am (MAINTAINERCLEANFILES):
Add the .properties and the .class files.
(EXTRA_DIST): Add the .properties files.
(update-properties, update-classes): Generate the .properties and .class files
in the po/ directory, not in the top-level directory.
* gettext-tools/examples/hello-java*/autoclean.sh: Remove the catalogs in
the po/ directory, not in the top-level directory.