intl: Fix grammar conflicts instead of accepting them.
The seven conflicts are related to "a ? b : c <op> d".  In all the
cases we left the S/R conflict be resolved as a shift, i.e., "a ? b
: (c <op> d)".  To make this explicit, give the rule of the ternary
operator a very low precedence.  In fact give it the same precedence
as "?".  Now, to resolve the conflict in "a ? b : c ? d : e" be parsed
as "a ? b : (c ? d : e)", make them right-associative.

* gettext-runtime/intl/plural.y: Require Bison 3.0, to use %precedence
instead of useless associativities such as %right.
Make '?' and ':' of the same precedence, right associative.